fix: comments and abstract utils
